+++ Sources/SwiftUIKit/Pickers/SimplePicker/Types+SimplePicker.swift
@@ -1,0 +1,14 @@
+//
+//  Types+SimplePicker.swift
+//  SwiftUIKit
+//
+//  Created by Daniel Saidi on 2021-05-09.
+//  Copyright © 2021 Daniel Saidi. All rights reserved.
+//
+
+import Foundation
+
+/**
+ This makes `SystemColor` implement `SimplePickerValue`.
+ */
+extension SystemColor: SimplePickerValue {}