+++ Release Notes.md
@@ -7,7 +7,8 @@
 
 * `FolderMonitor` can be used to monitor changes to a folder on the file system.
 * `FolderObservable` can be used to observe changes to a folder on the file system.
-* `ProcessInfo` has a new `isSwiftUIPreview` extension that can be used to tell if code is running within a preview or not. 
+* `ProcessInfo` has a new `isSwiftUIPreview` extension that can be used to tell if code is running within a preview or not.
+* `SystemColor` is a new enum that makes it easy to use and pick built-in system `Color` values.
 * `TagTextField` is a text field that automatically slugifies text as you type.