+++ Release Notes.md
@@ -1,7 +1,7 @@
 # Release notes
 
 
-## 1.3.2
+## 1.4.0
 
 ### ✨ New features